server/requirements.txt
fastapi>=0.75.0
uvicorn>=0.17.0
celery>=5.2.0
redis>=4.3.0
python-multipart>=0.0.5
reportlab>=3.6.0
pillow>=9.0.0
pytest>=7.0.0
httpx>=0.22.0